(#5) Restructure constants and add constant file stubs for Scoring mechanisms
