+++ src/main/java/frc/robot/constants/subsystems/scoring/IndexerConstants.java
@@ -1,0 +1,5 @@
+package frc.robot.constants.subsystems.scoring;
+
+public final class IndexerConstants {
+    // TODO: Use JSONSync
+}
