+++ src/main/java/frc/robot/constants/subsystems/scoring/ShooterConstants.java
@@ -1,0 +1,5 @@
+package frc.robot.constants.subsystems.scoring;
+
+public final class ShooterConstants {
+    // TODO: Use JSONSync
+}
